requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

mlflow==3.6.0
pandas==2.3.3
PyYAML==6.0.3
scikit-learn==1.7.2
